+++ cypress.config.ts
@@ -10,7 +10,19 @@ export default defineConfig({
     // We've imported your old cypress plugins here.
     // You may want to clean this up later by importing these.
     setupNodeEvents(on, config) {
-      installLogsPrinter(on);
+      let options = {};
+      if(config.env.CYPRESS_LOG === 'log') {
+        options = {
+          printLogsToFile: 'always',
+          outputRoot: config.projectRoot+'/logs/',
+          specRoot: 'cypress/e2e',
+          outputTarget: {
+            'cypress-logs|txt': 'txt',
+            'cypress-logs|json': 'json'
+          }
+        }
+      }
+      installLogsPrinter(on, options);
       return require('./cypress/plugins/index.ts')(on, config);
     },
   },